2.1.Embedded Checkout Form{Your Stripe Public and Private Keys}
Click on the documentation link at the top, and then click Embedded Form. There are two ways to build a checkout-form: the easy & lazy way - via an embedded form that Stripe builds for you - or the harder way - with an HTML form that you build yourself. Our sheep investors want us to hoof-it and get this live, so let's do the easy way first - and switch to a custom form later.

Getting the Form Script Code
To get the embedded form, copy the form code. Then, head to the app and open the templates/views/order/checkout.html.twig file. This is the template for the checkout page.
At the bottom, I already have a spot waiting for the checkout form. Paste the code there.
Oh! And as promised: this pk_test value is the public key from our test environment.

Your Stripe Public and Private Keys
Let me show you what I mean: in Stripe, open your "Account Settings" on the top and then select "API Keys". Each environment - Test and Live - have their own two keys: the secret key and the publishable or public key. Right now, we're using the public key for the test environment - so once we get this going, orders will show up there. After we deploy, we'll need to switch to the Live environment keys.
Oh, and, I think it's obvious - but these secret keys need to be kept secret. The last thing you should do is create a screencast and publish them to the web. Oh crap.
To resolve that create .env.local file. Write in this file:
STRIPE_PUBLIC_KEY=your_public_key
STRIPE_SECRET_KEY=your_secret_key
This file wouldn't be commited to repository so your real values are hidden for public eyes.
In services.yaml they can ce accessed with '%env(value)%' construct in parameters section. Then pass the parameter that retrieve public stripe key to checkout action in order controller and this value in twig checkout template.