2.3.Embedded Checkout Form{Fetching the Stripe Token}
Let's go get that token on the server. Open up src/Controller/OrderController.php and find checkoutAction().
This controller renders the checkout page. And because the HTML form has action="".
When Stripe submits the form, it submits right back to this same URL and controller.
To fetch the token, add a Request argument, and make sure you have the use statement on top.
Then, inside the method, say if ($request->isMethod('POST'), then we know the form was just submitted. If so, dump($request->get('stripeToken')).
If you read Stripe's documentation, that's the name of the hidden input field.
Try it out! Refresh and fill in the info again: use your trusty fake credit card number, some fake data and Pay. The form submits and the page refreshes. But thanks to the dump() function, hover over the target icon in the web debug toolbar. Perfect! We were able to fetch the token.
In a second, we're going to send this back to Stripe and ask them to actually charge the credit card it represents. But before we do that, head back to the Stripe dashboard.
Stripe shows you a log of pretty much everything that happens. Click the Logs link: these are all the interactions we've had with Stripe's API, including a few from before I hit record on the screencast. Click the first one: this is the AJAX request that the JavaScript made to Stripe: it sent over the credit card information, and Stripe sent back the token. If I search for the token that was just dumped, it matches.
Ok, let's use that token to charge our customer.