5.1.Stripe Customers + Our Users{Storing stripeUserId on the User Table}
As you can see, there are a lot of objects that you can interact with in Stripe's API. We've only talked about one: Charge. And it's enough to collect money.
But to create a really nice system, you need to talk about the Customer object. Customers give you three big super-powers.
First, if the same customer comes back over-and-over again, right now, they need to enter their credit card information every time. But with a Customer, you can store cards in Stripe and charge them using that. Second, all the charges for a customer will show up in one spot in Stripe's dashboard, instead of all over the place. And third, customers are needed to process subscription payments - something we'll talk about in the next Stripe course.

Storing stripeUserId on the User Table
So here's the goal: instead of creating random Charge objects, let's create a Customer object in Stripe and then charge that Customer. We also need to save the customer id to our user table so that when that user comes back in the future, we'll know that they are already a customer in Stripe.
In this project, the name of the user table is user, and it contains just some basic fields, like email, username and few others related to things like resetting your password.
Let's add a new column called stripeUserId. To do that, open a class in src\Entity called User. Create a new private property called stripeCustomerId
Above that, we're going to use annotations to say @Orm\Column(type="string") to create a varchar column. Let's even add a unique index on this field and add nullable=true to allow the field to be empty in the database.
At the bottom of the class, I'll use the Code->Generate menu - or Command+N on a Mac - to generate the getters and setters for the new property.