6.1.We <3 Creating Stripe Customers{Fetching the Existing Customer Object}
Head back to OrderController. Create a $user variable set to $this->getUser().
This is the User object for who is currently logged in. I'll add some inline documentation to show that.
When the user submits the payment form, there are two different scenarios. First, if (! $user->getStripeCustomerId()), then this is a first-time buyer, and we need to create a new Stripe Customer for them.
To do that, go back to their API documentation and find Create A Customer. Oh hey, it wrote the code for us again! Steal it! And paste it right inside the if statement. Customer has a lot of fields, but most are optional. Let's set email to $user->getEmail().
So we can easily look up a user in Stripe's dashboard later.
The really important field is source. This refers to the payment source - so credit or debit card in our case - that you want to attach to the customer. Set this to the $token variable.
This is huge: it will attach that card to their account, and allow us - if we want to - to charge them using that same card in the future.
Set this call to a new $customer variable: the create() method returns a Stripe\Customer object.
And we like that because this object has an id property.
To save that on our user record, say ->setStripeCustomerId($customer->id)
Then, I'll use Doctrine to run the UPDATE query to the database.
If you're not using Doctrine, just make sure to update the user record in the database however you want.

Fetching the Existing Customer Object
Now, add the else: this means the user already has a Stripe customer object. Repeat customer! Instead of creating a new one, just fetch the customer with \Stripe\Customer::retrieve() and pass it $user->getStripeCustomerId()