6.2.We <3 Creating Stripe Customers{Fetching the Existing Customer Object}
Since this user is already in Stripe, we might eventually re-work our checkout page so that they don't need to re-enter their credit card. But, we haven't done that yet. And since they just submitted fresh card information, we should update their account with that. After all, this might be a different card than what they used the first time they ordered.
To do that, update the source field: set it to $token. To send that update to Stripe, call $customer->save()
So in both situations, the token will now be attached to the customer that's associated with our user. Phew!