7.2.Stripe Invoices{One InvoiceItem per Product}
One InvoiceItem per Product
In OrderController, around the InvoiceItem part, add a foreach, over $this->cart->getProducts() as $product. In other words, let's loop over all the actual products in our cart and create a separate InvoiceItem for each. All we need to do is change the amount to be: $product->getPrice() * 100. We can even improve the description: set it to $product->getName()
Now, if we eventually send the user a receipt, it's going to be very easy to look on this Stripe invoice and see exactly what we charged them for.
Test time! Put in our awesome fake information, hit ENTER... and no errors.
In Stripe, click back to the customer page and find the new invoice - for $106. Click on that. Yes! 2 crystal clear invoice line items.
So yes, you can just charge customers. But if you create an Invoice with detailed line items, you're going to have a much better accounting system in the long run.