8.1.Centralize your Stripe Code{Moving createCustomer()}
Stripe's API is really organized. Our code that talks to it is getting a little crazy, unless you like long, procedural code that you can't re-use. Please tell me that's not the case.
Let's get this organized! At the very least, we should do this because eventually we're going to need to re-use some of this logic - particularly with subscriptions.
Here's the goal of the next few minutes: move each thing we're doing in the controller into a set of nice, re-usable functions. To do that, inside src/Stripe, create a new class called StripeClient.
Make sure this has the App\StripeClient namespace. We're going to fill this with functions that work with Stripe, like createCustomer() or updateCustomerCard().

Moving createCustomer()
In the controller, the first thing we do is create a Customer.
In StripeClient, add a new createCustomer() method that will accept the User object which should be associated with the customer, and the $paymentToken that was just submitted.
Copy the logic from the controller and paste it here. Update $token to $paymentToken. Then, return the $customer at the bottom, just in case we need it.
You'll see me do with this most functions in this class.
The only problem is with the entity manager - the code used to update the user record in the database. The way we fix this is a bit specific to Symfony. First, add a public function __construct() with an EntityManager $em argument. Set this on a new $em property.
Down below, just say $em = $this->em
In the controller, clear out all the code in the if statement.
This will be an instance of that StripeClient class.
In this if, call $stripeClient->createCustomer() and pass it the $user object and the $token.
Done.