8.2.Centralize your Stripe Code{Moving createCustomer()}
Moving updateCustomerCard()
Let's keep going!

The second piece of logic is responsible for updating the card on an existing customer. In StripeClient, add a public function updateCustomerCard() with a User $user whose related Customer should be updated, and the new $paymentToken.
Copy logic from the controller and past it here. Update $token to $paymentToken.
Go copy the logic from the controller, and paste it here. Update $token to $paymentToken.
In OrderController, call this with $stripeClient->updateCustomerCard() passing it $user and $token.
Now the StripeClient class is getting dangerous!
But, there's one small problem. This will work now, but look at the setApiKey() method call that's above everything:

src/Controller/OrderController.php
class OrderController extends extends AbstractController
{
    public function checkoutAction(Request $request)
    {
        if ($request->isMethod('POST')) {
            \Stripe\Stripe::setApiKey($this->getParameter('stripe_secret_key'));
        }
    }
}
We must call this before we make any API calls to Stripe. So, if we tried to use the StripeClient somewhere else in our code, but we forgot to call this line, we would have big problems.