8.3.Centralize your Stripe Code{Always setting the API Key}
Always setting the API Key
But, there's one small problem. This will work now, but look at the setApiKey() method call that's above everything.  I want to guarantee that if somebody calls a method on StripeClient, setApiKey() will always be called first. To do that, copy that line, delete it and move it into StripeClient's __construct() method.
Symfony user's will know that the getParameter() method won't work here. To fix that, add a new first constructor argument called $stripeSecretKey. Then, use that
To tell Symfony to pass this, go back to services.yaml and add a bind entry.
The end-result is this: when our StripeClient object is created, the API key is set immediately.