8.4.Centralize your Stripe Code{Moving Invoice Logic}
Moving Invoice Logic

Ok, the hard stuff is behind us: let's move the last two pieces of logic: creating an InvoiceItem and creating an Invoice. In StripeClient, add public function createInvoiceItem() with an $amount argument, the $user to attach it to and a $description.
Copy that code from our controller, remove it, and paste it here. Update amount to use $amount and description to use $description. Add a return statement just in case.
In OrderController, call this $stripeClient->createInvoiceItem() passing it $product->getPrice() * 100, $user and $product->getName()