8.5.Centralize your Stripe Code{Moving Invoice Logic}
Perfect! For the last piece, add a new public function createInvoice() with a $user whose customer we should invoice and a $payImmediately argument that defaults to true.
Who knows, there might be some time in the future when we don't want to pay an invoice immediately.
You know the drill: copy the invoice code from the controller, remove it and paste it into StripeClient. Wrap the pay() method inside if ($payImmediately). Finally, return the $invoice.
Call that in the controller: $stripeClient->createInvoice() passing it $user and true to pay immediately.
Phew! This was a giant step sideways - but not only is our code more re-usable, it just makes a lot more sense when you read it!
Double-check to make sure it works. Add something to your cart. Check-out. Yes! No error! The system still works and this StripeClient is really, really sweet.