10.1.Building the Custom Checkout Form{Setting up the Stripe JavaScript}
Earlier, we were rushing to get the site up and the sheep shopping. That's why we used Stripe's pre-built embedded form. And this is completely fine if you like it. But I want to build a custom form that looks like native on our site.
To do that, go back to the Stripe docs. Instead of embedded form, click "Custom Form". Using a custom form is very similar: we still send the credit card information to Stripe, and Stripe will still give us back a token. The difference is that we are responsible for building the HTML form.

Setting up the Stripe JavaScript
To help communicate with Stripe, we need some JavaScript. Copy the first JavaScript code and then find the checkout.html.twig template. At the top, override {% block javascripts %} and then call the {{ parent() }} function. Paste the script tag below.
This is just the Twig way of adding some new JavaScript to our page. The base layout also has a javascripts block and jQuery is already included.
Next, we need to tell the JavaScript about our publishable key. Copy that code from the docs and add it in the block.
We already know from our original code that we have a variable called stripe_public_key. Inside of the JavaScript quotes, print stripe_public_key.
Awesome!