11.1.Checkout Form JS Handling Logic
Now that we've got the form in place, we need to add some JavaScript that will send all the card information to Stripe. Once again, Stripe wrote a lot of this code for us. Over-achiever.
Copy their JavaScript and scroll up and paste it with our other JavaScript.
Then, back on the docs, scroll down a little further to another function called stripeResponseHandler(). Copy that too and paste it.
Let's look at the code: it uses a jQuery document.ready block to find the form and attach an on submit handler function:

 89 lines  templates/views/order/checkout.html.twig
{% block javascripts %}
    <script type="text/javascript">
        $(function () {
            var $form = $('#payment-form');
            $form.submit(function (event) {
            });
        });
    </script>
{% endblock %}
Because basically, when the user submits the form, we don't want to submit the form! Ahem, I mean, we want to stop, send all the information to Stripe, wait for the token to come back, put that in the form, and then submit it.