11.2.Checkout Form JS Handling Logic{Prepping the JS}
Prepping the JS

In our case, I've given the form a class called js-checkout-form. Copy that class, and change the JavaScript to look for .js-checkout-form. This is referenced in one other spot further below. Update that too.
It's not the most organized JS code.
Oh, and you'll notice that I use these js- classes a lot in my html. That's a standard that I like to use whenever I give an element a class not because I want to style it, but because I want to find it with JavaScript.
When this form is submitted, add event.preventDefault() to prevent the form from actually submitting.
This does more-or-less the same thing as returning false at the end of the function, but with some subtle differences.
Oof - let me fix some of this bad indentation. Next, the code finds the submit button so it can disable it. In our form, the button has a js-submit-button class. Copy that and update the code here, and once more down below.