11.3.Checkout Form JS Handling Logic{Fetching and Using the Token}
Fetching and Using the Token
Finally, here is the meat of the code. When we call Stripe.card.createToken():

templates/views/order/checkout.html.twig
{% block javascripts %}
    <script type="text/javascript">
        $(function () {
            $form.submit(function (event) {
                // Request a token from Stripe:
                Stripe.card.createToken($form, stripeResponseHandler);
            });
        });
    </script>
{% endblock %}
Stripe's Javascript will automatically fetch all the credit card data by reading the data-stripe attributes. Then, it sends those to stripe via AJAX. When that call finishes, it will execute the stripeResponseHandler function, and hopefully the response will contain that all-important token:

templates/views/order/checkout.html.twig
{% block javascripts %}
    <script type="text/javascript">
        function stripeResponseHandler(status, response) {
        }
    </script>
{% endblock %}
Now, if there was a problem with that card - like an invalid expiration - we need to show that error to the user. To do that, it looks for a payment-errors class and puts the message inside of that. I have a div ready for this. Its class is js-checkout-error and its hidden by default. Change the selector to .js-checkout-error, set the text, but then also call removeClass('hidden') so the element shows up.
Below in the else, life is good! I'll paste the .js-checkout-error code from before and modify it to re-add the hidden class - since now things are successful. When things work, the response comes back with a token, which we get via response.id. To send this to the server, we just smash it into a new input hidden field called... drumroll ... stripeToken.

This is precisely what the embedded form did. Once the form is submitted, the controller will hum along like nothing ever changed.

Testing the Error and Success
But, that's assuming we didn't mess something up! That's a big but. Go back and refresh the page.
First, test that the error handling works by adding an expiration date in the past. Put in the real credit card number -- oof, ugly formatting - we'll fix that. Then, use an expired expiration. Hit checkout and... boom!
It sent the info to Stripe, Stripe came back with an error, we put the error in the box, and showed that box to our user. In other words, we're awesome. Change this to a future expiration and try again.
It's alive!!!
The only problem I can think of now is how ugly entering a credit card number is: all those numbers just run together. The expiration field is a mess too. Oof. Let's fix that - it's surprisingly easy!